LPA DONE

+++ includes/lexer_parser_analyzer.h
@@ -6,7 +6,7 @@
 /*   By: arsciand <arsciand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/11 16:53:44 by arsciand          #+#    #+#             */
-/*   Updated: 2020/03/08 15:26:38 by arsciand         ###   ########.fr       */
+/*   Updated: 2020/03/08 20:59:34 by arsciand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,7 +49,7 @@ void		free_process_list(t_lst **head);
 void		ft_free_redirlist(t_lst **head);
 char		*fill_cmd_job(t_lst *tok_lst, char *cmd);
 char		**ft_add_arg_cmd_process(char **tablo, const char *str);
-char		*join_one_char(char *dst, char *str, int index);
+char		*join_one_char(char *dst, char *str, size_t index);
 
 /*
 **	PARSER
